# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AllAmericanEnergy CRM API
  version: 0.1.0
servers:
  - url: http://localhost:4000/api
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
  /auth/login:
    post:
      summary: Login with email and password
      responses:
        "200":
          description: Access and refresh tokens
  /auth/refresh:
    post:
      summary: Refresh an access token
      responses:
        "200":
          description: New access token
  /auth/logout:
    post:
      summary: Revoke a refresh token
      responses:
        "204":
          description: Token revoked
  /users/me:
    get:
      summary: Current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile and permissions
  /companies:
    get:
      summary: List companies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Paginated companies
    post:
      summary: Create company
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Company created
  /contacts:
    get:
      summary: List contacts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Paginated contacts
    post:
      summary: Create contact
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Contact created
  /deals:
    get:
      summary: List deals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Paginated deals
    post:
      summary: Create deal
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Deal created
  /tasks:
    get:
      summary: List tasks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Paginated tasks
    post:
      summary: Create task
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Task created
  /reports/pipeline:
    get:
      summary: Pipeline report
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pipeline KPIs by stage
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT